refactor(v0.70.0): scheduler.py 1208L → 9-module 패키지로 분할 (Tier 3 #2)
Tier 3 God Object 분할 시리즈 두 번째. 7 class + 14 헬퍼가 한 파일에
묶여 있던 1208 LOC monolith를 관심사별 9개 모듈로 분리.

새 패키지 구조 — core/scheduler/scheduler/:
  __init__.py       81L  24개 심볼 re-export (외부 11곳 import 변경 0)
  models.py         93L  ScheduleKind/Schedule/ActiveHours/ScheduledJob + 6 상수 + OnJobFired
  serialization.py  81L  _job_to_dict, _job_from_dict
  run_log.py        93L  JobRunLog (per-job JSONL + auto-pruning)
  lock.py          135L  SchedulerLock + _is_pid_alive (_try_reclaim 의존)
  jitter.py         38L  _compute_jitter_frac, _jittered_next_run
  timezone.py       59L  _parse_hhmm, _now_minutes, _cron_tuple_for_tz
  service.py       708L  SchedulerService 본체 (의도적 단일 보존)
  factory.py        42L  create_scheduler

공개 API 변경 0건. 외부 import 11곳 (core/ 3, tests/ 8) 수정 불필요 —
패키지 __init__.py가 SchedulerService, Schedule, ScheduledJob,
ScheduleKind, ActiveHours, JobRunLog, SchedulerLock, _is_pid_alive,
_compute_jitter_frac, _jittered_next_run, _job_to_dict, _job_from_dict,
_parse_hhmm, _now_minutes, _cron_tuple_for_tz, MISSED_TASK_GRACE_MS,
RECURRING_MAX_AGE_MS, DEFAULT_*, OnJobFired, create_scheduler 등 24개
심볼 모두 re-export.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (268 source files, +8)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- 스케줄러 전용 테스트 245개 모두 통과
- E2E Cowboy Bebop A (68.4) 변동 없음
- 최대 단일 파일 1208 → 708 LOC (-41%)

남은 Tier 3 후보 7개 — commands.py(2441), cli/__init__.py(1889),
agent/loop.py(1754), agentic_ui.py(1160), reports.py(1156),
tool_executor.py(1047), llm/router.py(1046).
