+++ core/scheduler/scheduler/service.py
@@ -1,514 +1,44 @@
-"""Advanced Scheduler — 3-type scheduling + active hours.
+"""SchedulerService — the central scheduling engine.
 
-Inspired by OpenClaw's Cron Service + Claude Code's distributed scheduler.
-Supports:
-- AT: one-shot absolute timestamp jobs
-- EVERY: fixed-interval with anchor-based drift prevention
-- CRON: standard cron expressions (via existing CronParser)
-- Active Hours: timezone-aware quiet-hours window with midnight wrap-around
-- Per-job JSONL run log with auto-pruning
-- Atomic JSON store (tmp + rename pattern)
-- O_EXCL lock file + PID liveness probe (claude-code pattern)
-- Deterministic per-job jitter (thundering herd prevention)
-- Session-only tasks (durable flag)
-- Missed task recovery
-- mtime-based file watch reload (1s check interval)
-
-Architecture-v6 SS4.5: Automation Layer -- Advanced Scheduler (P4).
+Combines the data models, run log, lock, jitter, and timezone helpers from
+sibling modules into a single long-running service that ticks once per
+second and fires due jobs.
 """
 
 from __future__ import annotations
 
 import contextlib
-import hashlib
 import json
 import logging
 import os
 import threading
 import time
-from collections.abc import Callable
-from dataclasses import asdict, dataclass, field
-from enum import Enum
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
-from core.paths import PROJECT_SCHEDULER_FILE, PROJECT_SCHEDULER_LOG_DIR
+from core.scheduler.scheduler.jitter import _jittered_next_run
+from core.scheduler.scheduler.lock import SchedulerLock
+from core.scheduler.scheduler.models import (
+    _LEGACY_STORE_PATH,
+    DEFAULT_LOG_DIR,
+    DEFAULT_MAX_JITTER_MS,
+    DEFAULT_STORE_PATH,
+    MISSED_TASK_GRACE_MS,
+    RECURRING_MAX_AGE_MS,
+    ActiveHours,
+    OnJobFired,
+    ScheduledJob,
+    ScheduleKind,
+)
+from core.scheduler.scheduler.run_log import JobRunLog
+from core.scheduler.scheduler.serialization import _job_from_dict, _job_to_dict
+from core.scheduler.scheduler.timezone import _cron_tuple_for_tz, _now_minutes, _parse_hhmm
 from core.scheduler.triggers import CronParser, TriggerManager
 
 if TYPE_CHECKING:
     from core.hooks import HookSystem
 
 log = logging.getLogger(__name__)
-
-# ---------------------------------------------------------------------------
-# Type aliases
-# ---------------------------------------------------------------------------
-
-OnJobFired = Callable[[str, str, bool, str], None]
-"""Callback: (job_id, action, isolated, agent_id) -> None."""
-
-# ---------------------------------------------------------------------------
-# Constants
-# ---------------------------------------------------------------------------
-
-DEFAULT_STORE_PATH = PROJECT_SCHEDULER_FILE
-DEFAULT_LOG_DIR = PROJECT_SCHEDULER_LOG_DIR
-
-# Backward-compat: old global store path for migration detection
-_LEGACY_STORE_PATH = Path.home() / ".geode" / "scheduler" / "jobs.json"
-
-# Missed task recovery: 1 hour grace window
-MISSED_TASK_GRACE_MS: float = 3_600_000.0
-
-# Jitter defaults
-DEFAULT_MAX_JITTER_MS: float = 900_000.0  # 15 minutes cap
-DEFAULT_JITTER_FRACTION: float = 0.1  # 10% of interval
-
-# Recurring age-out: 30 days (claude-code pattern)
-RECURRING_MAX_AGE_MS: float = 30 * 24 * 60 * 60 * 1000.0
-
-
-# ---------------------------------------------------------------------------
-# Data models
-# ---------------------------------------------------------------------------
-
-
-class ScheduleKind(Enum):
-    """Supported schedule types."""
-
-    AT = "at"  # One-shot absolute timestamp
-    EVERY = "every"  # Fixed interval with anchor
-    CRON = "cron"  # Cron expression
-
-
-@dataclass
-class Schedule:
-    """Schedule configuration for a job."""
-
-    kind: ScheduleKind
-    at_ms: float = 0.0  # For AT: Unix timestamp in ms
-    every_ms: float = 0.0  # For EVERY: interval in ms
-    anchor_ms: float = 0.0  # For EVERY: anchor time for drift prevention
-    cron_expr: str = ""  # For CRON: cron expression
-    timezone: str = ""  # For CRON/active hours: IANA timezone
-
-
-@dataclass
-class ActiveHours:
-    """Quiet-hours window configuration.
-
-    When set, jobs only run within the specified time window.
-    Supports midnight wrap-around (e.g., start="22:00", end="06:00").
-    """
-
-    start: str = ""  # "HH:MM" format
-    end: str = ""  # "HH:MM" format
-    timezone: str = ""  # IANA timezone or "" for local
-
-
-@dataclass
-class ScheduledJob:
-    """A job managed by the SchedulerService."""
-
-    job_id: str
-    name: str
-    schedule: Schedule
-    enabled: bool = True
-    delete_after_run: bool = False  # For AT type: auto-delete after success
-    durable: bool = True  # False = session-only (not persisted to disk)
-    permanent: bool = False  # Exempt from recurring age-out
-    agent_id: str = ""  # Sub-agent owner (empty = main session)
-    callback: Any = None  # Callable[[dict], None]
-    action: str = ""  # Prompt text to enqueue when fired (no callback)
-    isolated: bool = True  # Run in isolated session (agentTurn) vs main session (systemEvent)
-    active_hours: ActiveHours | None = None
-    metadata: dict[str, Any] = field(default_factory=dict)
-    # State tracking
-    created_at_ms: float = 0.0
-    next_run_at_ms: float | None = None
-    last_run_at_ms: float | None = None
-    last_status: str = ""  # "ok" | "error" | "skipped" | "stuck"
-    last_duration_ms: float = 0.0
-    running_since_ms: float | None = None  # Set when executing, cleared on completion
-
-
-# ---------------------------------------------------------------------------
-# Serialisation helpers
-# ---------------------------------------------------------------------------
-
-
-def _job_to_dict(job: ScheduledJob) -> dict[str, Any]:
-    """Serialize a ScheduledJob to a JSON-safe dict (excludes callback)."""
-    return {
-        "job_id": job.job_id,
-        "name": job.name,
-        "schedule": {
-            "kind": job.schedule.kind.value,
-            "at_ms": job.schedule.at_ms,
-            "every_ms": job.schedule.every_ms,
-            "anchor_ms": job.schedule.anchor_ms,
-            "cron_expr": job.schedule.cron_expr,
-            "timezone": job.schedule.timezone,
-        },
-        "enabled": job.enabled,
-        "delete_after_run": job.delete_after_run,
-        "durable": job.durable,
-        "permanent": job.permanent,
-        "agent_id": job.agent_id,
-        "action": job.action,
-        "isolated": job.isolated,
-        "active_hours": (asdict(job.active_hours) if job.active_hours is not None else None),
-        "metadata": job.metadata,
-        "created_at_ms": job.created_at_ms,
-        "next_run_at_ms": job.next_run_at_ms,
-        "last_run_at_ms": job.last_run_at_ms,
-        "last_status": job.last_status,
-        "last_duration_ms": job.last_duration_ms,
-        "running_since_ms": job.running_since_ms,
-    }
-
-
-def _job_from_dict(data: dict[str, Any]) -> ScheduledJob:
-    """Deserialize a dict into a ScheduledJob (callback will be None)."""
-    sched_data = data["schedule"]
-    schedule = Schedule(
-        kind=ScheduleKind(sched_data["kind"]),
-        at_ms=sched_data.get("at_ms", 0.0),
-        every_ms=sched_data.get("every_ms", 0.0),
-        anchor_ms=sched_data.get("anchor_ms", 0.0),
-        cron_expr=sched_data.get("cron_expr", ""),
-        timezone=sched_data.get("timezone", ""),
-    )
-    ah_data = data.get("active_hours")
-    active_hours = ActiveHours(**ah_data) if ah_data is not None else None
-
-    return ScheduledJob(
-        job_id=data["job_id"],
-        name=data["name"],
-        schedule=schedule,
-        enabled=data.get("enabled", True),
-        delete_after_run=data.get("delete_after_run", False),
-        durable=data.get("durable", True),
-        permanent=data.get("permanent", False),
-        agent_id=data.get("agent_id", ""),
-        callback=None,  # Callbacks are not serialised
-        action=data.get("action", ""),
-        isolated=data.get("isolated", True),
-        active_hours=active_hours,
-        metadata=data.get("metadata", {}),
-        created_at_ms=data.get("created_at_ms", 0.0),
-        next_run_at_ms=data.get("next_run_at_ms"),
-        last_run_at_ms=data.get("last_run_at_ms"),
-        last_status=data.get("last_status", ""),
-        last_duration_ms=data.get("last_duration_ms", 0.0),
-        running_since_ms=data.get("running_since_ms"),
-    )
-
-
-# ---------------------------------------------------------------------------
-# Per-job JSONL run log
-# ---------------------------------------------------------------------------
-
-
-class JobRunLog:
-    """Per-job JSONL run log with auto-pruning.
-
-    Each job gets its own ``{job_id}.jsonl`` file under *log_dir*.
-    """
-
-    MAX_LINES: int = 2000
-    MAX_BYTES: int = 2 * 1024 * 1024  # 2 MB
-
-    def __init__(self, log_dir: Path | None = None) -> None:
-        self._log_dir = log_dir if log_dir is not None else DEFAULT_LOG_DIR
-        self._lock = threading.Lock()
-
-    def _path(self, job_id: str) -> Path:
-        safe = job_id.replace(":", "_").replace("/", "_")
-        return self._log_dir / f"{safe}.jsonl"
-
-    def append(self, job_id: str, entry: dict[str, Any]) -> None:
-        """Append a run entry for *job_id*."""
-        path = self._path(job_id)
-        path.parent.mkdir(parents=True, exist_ok=True)
-        line = json.dumps(entry, ensure_ascii=False, separators=(",", ":"))
-        with self._lock, open(path, "a", encoding="utf-8") as f:
-            f.write(line + "\n")
-
-    def get_runs(self, job_id: str, limit: int = 100) -> list[dict[str, Any]]:
-        """Read the most recent *limit* entries (newest first)."""
-        path = self._path(job_id)
-        if not path.exists():
-            return []
-        with open(path, encoding="utf-8") as f:
-            lines = f.readlines()
-        entries: list[dict[str, Any]] = []
-        for raw in reversed(lines):
-            raw = raw.strip()
-            if not raw:
-                continue
-            try:
-                entries.append(json.loads(raw))
-            except json.JSONDecodeError:
-                log.warning("Skipping malformed run log line: %s", raw[:80])
-            if len(entries) >= limit:
-                break
-        return entries
-
-    def prune(self, job_id: str) -> int:
-        """Prune log file if it exceeds size/line limits.
-
-        Returns the number of lines removed.
-        """
-        with self._lock:
-            path = self._path(job_id)
-            if not path.exists():
-                return 0
-            file_size = path.stat().st_size
-            if file_size <= self.MAX_BYTES:
-                return 0
-            with open(path, encoding="utf-8") as f:
-                lines = f.readlines()
-            original = len(lines)
-            if original <= self.MAX_LINES:
-                return 0
-            kept = lines[-self.MAX_LINES :]
-            removed = original - len(kept)
-            # Atomic write: tmp + rename
-            tmp_path = path.with_suffix(".jsonl.tmp")
-            with open(tmp_path, "w", encoding="utf-8") as f:
-                f.writelines(kept)
-            os.replace(str(tmp_path), str(path))
-            log.info(
-                "Pruned job run log %s: %d -> %d lines (%d removed)",
-                job_id,
-                original,
-                len(kept),
-                removed,
-            )
-            return removed
-
-
-# ---------------------------------------------------------------------------
-# O_EXCL Lock — PID liveness probe (claude-code pattern)
-# ---------------------------------------------------------------------------
-
-
-class SchedulerLock:
-    """Cross-platform file-based lock using O_EXCL atomic creation.
-
-    Lock file contains ``{"pid": N, "acquired_at": T, "session_id": S}``
-    for liveness probing and idempotent re-acquire on session restart.
-    """
-
-    STALE_TIMEOUT_S: float = 5.0
-
-    def __init__(self, lock_path: Path, session_id: str = "") -> None:
-        self._lock_path = lock_path
-        self._session_id = session_id or f"pid-{os.getpid()}"
-        self._acquired = False
-
-    def _lock_content(self) -> str:
-        return json.dumps(
-            {
-                "pid": os.getpid(),
-                "acquired_at": time.time(),
-                "session_id": self._session_id,
-            }
-        )
-
-    def acquire(self, timeout_s: float = 10.0) -> bool:
-        """Try to acquire the lock, probing for stale owners."""
-        deadline = time.monotonic() + timeout_s
-        while time.monotonic() < deadline:
-            try:
-                fd = os.open(
-                    str(self._lock_path),
-                    os.O_CREAT | os.O_EXCL | os.O_WRONLY,
-                    0o644,
-                )
-                try:
-                    os.write(fd, self._lock_content().encode())
-                finally:
-                    os.close(fd)
-                self._acquired = True
-                return True
-            except FileExistsError:
-                # Lock file exists — check if owner is alive or same session
-                reclaim_result = self._try_reclaim()
-                if reclaim_result == "idempotent":
-                    # Same session re-acquired — lock file updated in place
-                    self._acquired = True
-                    return True
-                if reclaim_result == "reclaimed":
-                    continue  # Stale lock removed, retry O_EXCL
-                time.sleep(0.2)
-        return False
-
-    def release(self) -> None:
-        """Release the lock by removing the lock file."""
-        if self._acquired:
-            with contextlib.suppress(OSError):
-                self._lock_path.unlink(missing_ok=True)
-            self._acquired = False
-
-    def _try_reclaim(self) -> str:
-        """Check if the current lock owner is dead and reclaim if so.
-
-        Returns:
-            "idempotent" — same session re-acquired (lock file updated in place)
-            "reclaimed"  — stale lock removed (caller should retry O_EXCL)
-            ""           — lock is held by another live session
-        """
-        try:
-            with open(self._lock_path, encoding="utf-8") as f:
-                data = json.load(f)
-            pid = data.get("pid", -1)
-            acquired_at = data.get("acquired_at", 0.0)
-            owner_session = data.get("session_id", "")
-            age = time.time() - acquired_at
-
-            # Idempotent: same session, new PID (restart/resume)
-            if owner_session and owner_session == self._session_id:
-                try:
-                    self._lock_path.write_text(self._lock_content())
-                    return "idempotent"
-                except OSError:
-                    return ""
-
-            if not _is_pid_alive(pid) or age > 600:
-                # Owner dead or lock very old — reclaim
-                try:
-                    self._lock_path.unlink()
-                    return "reclaimed"
-                except FileNotFoundError:
-                    return "reclaimed"
-                except OSError:
-                    return ""
-        except (json.JSONDecodeError, OSError):
-            # Corrupt lock file — remove and retry
-            try:
-                self._lock_path.unlink()
-                return "reclaimed"
-            except OSError:
-                return ""
-        return ""
-
-    def __enter__(self) -> SchedulerLock:
-        if not self.acquire():
-            raise TimeoutError(f"Could not acquire scheduler lock: {self._lock_path}")
-        return self
-
-    def __exit__(self, *_args: Any) -> None:
-        self.release()
-
-
-def _is_pid_alive(pid: int) -> bool:
-    """Check if a process with the given PID is alive."""
-    if pid <= 0:
-        return False
-    try:
-        os.kill(pid, 0)
-        return True
-    except ProcessLookupError:
-        return False
-    except PermissionError:
-        return True  # Process exists but we can't signal it
-
-
-# ---------------------------------------------------------------------------
-# Jitter — deterministic per-job hash (claude-code pattern)
-# ---------------------------------------------------------------------------
-
-
-def _compute_jitter_frac(job_id: str) -> float:
-    """Compute a deterministic jitter fraction [0, 1) from job_id.
-
-    Uses SHA-256 to produce a stable, uniformly distributed value.
-    Same job_id always produces the same jitter across restarts.
-    """
-    h = hashlib.sha256(job_id.encode()).digest()
-    # Use first 4 bytes as uint32 → normalize to [0, 1)
-    val = int.from_bytes(h[:4], "big")
-    return val / (2**32)
-
-
-def _jittered_next_run(
-    next_ms: float,
-    interval_ms: float,
-    job_id: str,
-    *,
-    max_jitter_ms: float = DEFAULT_MAX_JITTER_MS,
-    jitter_fraction: float = DEFAULT_JITTER_FRACTION,
-) -> float:
-    """Apply deterministic forward jitter to a recurring job's next-run time.
-
-    Jitter = min(frac * interval * jitter_fraction, max_jitter_ms).
-    This spreads jobs with the same cron across a window instead of all
-    firing at :00.
-    """
-    frac = _compute_jitter_frac(job_id)
-    jitter = min(frac * interval_ms * jitter_fraction, max_jitter_ms)
-    return next_ms + jitter
-
-
-# ---------------------------------------------------------------------------
-# Active-hours helpers
-# ---------------------------------------------------------------------------
-
-
-def _parse_hhmm(s: str) -> int:
-    """Parse "HH:MM" -> total minutes since midnight."""
-    parts = s.strip().split(":")
-    if len(parts) != 2:
-        raise ValueError(f"Invalid HH:MM format: '{s}'")
-    return int(parts[0]) * 60 + int(parts[1])
-
-
-def _now_minutes(timezone: str) -> int:
-    """Current local time as minutes since midnight.
-
-    When *timezone* is empty or not available, falls back to local time.
-    Uses ``datetime`` if *zoneinfo* is available for tz-aware resolution.
-    """
-    if timezone:
-        try:
-            import datetime
-            import zoneinfo
-
-            tz = zoneinfo.ZoneInfo(timezone)
-            now = datetime.datetime.now(tz=tz)
-            return now.hour * 60 + now.minute
-        except Exception:
-            log.debug("Timezone '%s' unavailable, falling back to local", timezone)
-    t = time.localtime()
-    return t.tm_hour * 60 + t.tm_min
-
-
-def _cron_tuple_for_tz(timezone: str) -> tuple[int, int, int, int, int]:
-    """Get a cron-compatible tuple, optionally in a specific timezone.
-
-    Returns weekday in standard cron convention: 0=Sun, 1=Mon, ..., 6=Sat.
-    Python's ``datetime.weekday()`` uses 0=Mon, so we convert:
-    ``(weekday() + 1) % 7``.
-    """
-    if timezone:
-        try:
-            import datetime
-            import zoneinfo
-
-            tz = zoneinfo.ZoneInfo(timezone)
-            now = datetime.datetime.now(tz=tz)
-            cron_wday = (now.weekday() + 1) % 7  # Python 0=Mon -> cron 0=Sun
-            return (now.minute, now.hour, now.day, now.month, cron_wday)
-        except Exception:  # noqa: S110 — fallback to CronParser.current_tuple() below
-            pass
-    return CronParser.current_tuple()
-
-
-# ---------------------------------------------------------------------------
-# Scheduler Service
-# ---------------------------------------------------------------------------
 
 
 class SchedulerService:
@@ -1174,35 +704,5 @@ class SchedulerService:
             self._stop_event.wait(interval_s)
 
 
-# ---------------------------------------------------------------------------
-# Factory function — claude-code-style library instantiation
-# ---------------------------------------------------------------------------
-
-
-def create_scheduler(
-    *,
-    store_path: Path | None = None,
-    log_dir: Path | None = None,
-    on_job_fired: OnJobFired | None = None,
-    trigger_manager: TriggerManager | None = None,
-    hooks: HookSystem | None = None,
-    session_id: str = "",
-    enable_jitter: bool = True,
-    max_jitter_ms: float = DEFAULT_MAX_JITTER_MS,
-) -> SchedulerService:
-    """Create a SchedulerService with project-local defaults.
-
-    This factory function makes the scheduler usable as a library component
-    that any context (serve, REPL, SDK, daemon) can instantiate with
-    appropriate configuration.
-    """
-    return SchedulerService(
-        trigger_manager=trigger_manager,
-        hooks=hooks,
-        store_path=store_path,
-        log_dir=log_dir,
-        on_job_fired=on_job_fired,
-        session_id=session_id,
-        enable_jitter=enable_jitter,
-        max_jitter_ms=max_jitter_ms,
-    )
+# Re-export DEFAULT_LOG_DIR for service consumers that need the constant
+__all__ = ["DEFAULT_LOG_DIR", "SchedulerService"]
